convert inro classes
